requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
Babel==2.11.0
certifi==2022.12.7
charset-normalizer==3.1.0
dump-env==1.4.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
Pillow==9.4.0
pydantic==1.10.5
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
